Testes

+++ src/main/java/com/saimon/FifoProject/Service/impl/TaskServiceImpl.java
@@ -3,6 +3,10 @@ package com.saimon.FifoProject.Service.impl;
 import com.saimon.FifoProject.Service.TaskService;
 import com.saimon.FifoProject.api.Model.Entity.MyTask;
 import com.saimon.FifoProject.api.Model.Repository.TaskRepository;
+import org.springframework.data.domain.Example;
+import org.springframework.data.domain.ExampleMatcher;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -27,12 +31,29 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public void delete(MyTask foundTask) {
+    public void delete(MyTask foundTask) throws IllegalAccessException {
+        if (foundTask == null || foundTask.getId() == null){
+            throw new IllegalAccessException("Task id can't be null.");
+        }
         taskRepository.delete(foundTask);
     }
 
     @Override
-    public MyTask update(MyTask foundTask) {
+    public MyTask update(MyTask foundTask) throws IllegalAccessException {
+        if (foundTask == null || foundTask.getId() == null){
+            throw new IllegalAccessException("Task id can't be null.");
+        }
         return taskRepository.save(foundTask);
+    }
+
+    @Override
+    public Page<MyTask> find(MyTask task, Pageable pageRequest) {
+        Example<MyTask> example = Example.of(task,
+                ExampleMatcher
+                        .matching()
+                        .withIgnoreCase()
+                        .withIgnoreNullValues()
+                        .withStringMatcher(ExampleMatcher.StringMatcher.CONTAINING));
+        return taskRepository.findAll(example, pageRequest);
     }
 }
